Initial commit of OpenSim Bot
This bot was created for a master's thesis project. It is designed
for use within an OpenSim server. The main purpose of the bot is
to extract a certain type of steganographic message from textures
which are found on a server. The bot was created using C# and the
libOpenMetaverse library.